Rename integrate to integration
