+++ imgal_core/src/integration/rectangle.rs
