+++ imgal_core/src/integration/simpson.rs
