+++ imgal_python/src/child_modules/integration_module.rs
@@ -4,26 +4,26 @@ use crate::functions::integration_functions;
 use crate::utils::py_import_module;
 
 /// Python binding for the "integrate" submodule.
-pub fn register_integrate_module(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
-    let integrate_module = PyModule::new(parent_module.py(), "integrate")?;
+pub fn register_integration_module(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
+    let integration_module = PyModule::new(parent_module.py(), "integrate")?;
 
     // add module to python's sys.modules
-    py_import_module("integrate");
+    py_import_module("integration");
 
     // add integrate submodule functions
-    integrate_module.add_function(wrap_pyfunction!(
+    integration_module.add_function(wrap_pyfunction!(
         integration_functions::integration_composite_simpson,
-        &integrate_module
+        &integration_module
     )?)?;
-    integrate_module.add_function(wrap_pyfunction!(
+    integration_module.add_function(wrap_pyfunction!(
         integration_functions::integration_midpoint,
-        &integrate_module
+        &integration_module
     )?)?;
-    integrate_module.add_function(wrap_pyfunction!(
+    integration_module.add_function(wrap_pyfunction!(
         integration_functions::integration_simpson,
-        &integrate_module
+        &integration_module
     )?)?;
 
     // attach to parent module
-    parent_module.add_submodule(&integrate_module)
+    parent_module.add_submodule(&integration_module)
 }
