+++ imgal_python/src/parent_module.rs
@@ -6,7 +6,7 @@ use super::child_modules::{integration_module, parameters_module, phasor_module,
 #[pymodule(name = "imgal_python")]
 fn imgal_parent_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
     // register child modules
-    integration_module::register_integrate_module(m)?;
+    integration_module::register_integration_module(m)?;
     parameters_module::register_parameters_module(m)?;
     phasor_module::register_phasor_module(m)?;
     simulation_module::register_simulation_module(m)?;
